[MOD] Rename scenarioAction to scenarioActions
