+++ src/reducer/scenario/scenarioActions.ts
